Used clang-tidy/-format on all source files
The source code was quite difficult to read as it lacked curly braces
around many if/else/for loops, and had a mix of tabs/spaces, not just
for the indents, but throughout the code/comments in general.

I used the clang tidy and format tools to reformat the source code,
with lots of manual work on top.

It was a laborious task, but will make life a lot easier going forward.
I just hope I didn't introduce any bugs during this refactoring!