Closes #1: Add .clang-format
The code has been reformatted to use the new configuration settings.

This config mostly follows the LLVM style with a few minor changes. Most
notably is the disabling of ColumnLimit.

Eventually I would like to set an 80 character column width, but having long
lines helped in the early refactorings, and I find it especially useful in
understanding the complexity in various parts of the code, such as those many
long if/else staements.